Remove logic from User, Wishlist controller

+++ app/src/main/java/com/future/medan/backend/services/impl/WishlistImpl.java
@@ -1,5 +1,6 @@
 package com.future.medan.backend.services.impl;
 
+import com.future.medan.backend.exceptions.ResourceNotFoundException;
 import com.future.medan.backend.models.entity.Wishlist;
 import com.future.medan.backend.repositories.WishlistRepository;
 import com.future.medan.backend.services.WishlistService;
@@ -7,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class WishlistImpl implements WishlistService {
@@ -25,8 +25,8 @@ public class WishlistImpl implements WishlistService {
     }
 
     @Override
-    public Optional<Wishlist> getById(String id){
-        return wishlistRepository.findById(id);
+    public Wishlist getById(String id){
+        return wishlistRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Wishlist", "id", id));
     }
 
     @Override
@@ -35,7 +35,18 @@ public class WishlistImpl implements WishlistService {
     }
 
     @Override
+    public Wishlist save(Wishlist wishlist, String id){
+        if (!wishlistRepository.existsById(id))
+            throw new ResourceNotFoundException("Wishlist", "id", id);
+        else
+            return wishlistRepository.save(wishlist);
+    }
+
+    @Override
     public void deleteById(String id){
-        wishlistRepository.deleteById(id);
+        if (!wishlistRepository.existsById(id))
+            throw new ResourceNotFoundException("Wishlist", "id", id);
+        else
+            wishlistRepository.deleteById(id);
     }
 }
